% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAS.r
\name{sas}
\alias{sas}
\title{Submits SAS file to console for execution}
\usage{
sas(SASFile)
}
\arguments{
\item{SASFile}{path to the SAS file, without extension}
}
\value{
validation test results
}
\description{
Submits SAS file to console for execution
}
